% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/literatureVault.R
\name{get_citations_pubmed}
\alias{get_citations_pubmed}
\title{A function that returns a citation with first author, journal and year for a PubMed ID}
\usage{
get_citations_pubmed(pmid, chunk_size = 100)
}
\arguments{
\item{pmid}{An array of Pubmed IDs}

\item{chunk_size}{Size of PMID chunks}
}
\value{
citation PubMed citation, with first author, journal and year
}
\description{
A function that returns a citation with first author, journal and year for a PubMed ID
}
